Implemented TheXaman's debug menu
The following have been added by me:
-Options to check a Pokémon's EVs and IVs
-Option to force hatch an Egg
-Option to give the Player an Egg
-Option to open the Pokémon Storage System

Also modified the DoPCTurnOnEffect and DoPCTurnOffEffect specials.
By default, they were creating a PC in the same tile the Player was standing still, so I made them check the tile in front of the Player in order to do their thing.
Date of creation: Mon May 3 22:45:24 2021 -0300