Made the game read uncompressed trainer sprites palettes Thanks to Josh/ShinyDragonHunter.
Originally, I did this to see if, for whatever reason, it solved the issue with the battle mugshots in Route 103. It didn't.
However, trustworthy sources of information (MGriffin and Josh themself) state that uncompressing the trainer palettes brings some benefits.
-Faster loading which is natural as the hardware doesn't need to do an extra step to uncompress them.
-Smaller palettes, as the compression actually makes them bigger.
-As a direct result, I get some (probably insignificant) amount of free space in the ROM for free.
Because of that, I'm keeping this in.
Date of creation: Sat Aug 28 11:29:05 2021 -0300